requirements.txt
flask==2.0.1
python-pptx==0.6.21
google-api-python-client==2.65.0
werkzeug==2.0.1